fix/artwork- image-flicker (#3)
* chore: update dependencies

* feat: artwork images and scripts

All the components which were previously utilizing temp images which were flickering because of the random assignment are now utilizing the image urls retrieved from cmoa's server.
The scripts for the images are added. There's a lot of reverse engineering that was done to grab those images. Some of the artworks are not able to retrieve the url because of a 403 error. More work is required to successfully get all the images.

* chore: schema format

* feat: local db

* fix: add margins to main component

* fix: remove alt, add aria label

* fix: image width

* fix: some data cleanups

---------

Co-authored-by: sbracellari <39267413+sbracellari@users.noreply.github.com>